Refactor: move RBA classes into their own file
Part of the ongoing breakup of binding.py
